refactor: merge deduplication tests into main InvitationManager spec
Add all deduplication-related tests under a `describe "handling deduplication"`
block in spec/services/invitation_manager_spec.rb, per review feedback on
PR #2551 (discussion_r3055854926). Delete the separate
invitation_manager_deduplication_spec.rb.
